+++ packages/cli/src/commands/page/index.ts
@@ -1,5 +1,6 @@
 import { defineCommand } from 'citty'
 
+import bounds from './bounds.ts'
 import current from './current.ts'
 import list from './list.ts'
 import set from './set.ts'
@@ -9,6 +10,7 @@ export default defineCommand({
   subCommands: {
     current,
     list,
-    set
+    set,
+    bounds
   }
 })