+++ packages/cli/src/commands/variable/index.ts
@@ -3,6 +3,7 @@ import { defineCommand } from 'citty'
 import bind from './bind.ts'
 import create from './create.ts'
 import deleteCmd from './delete.ts'
+import find from './find.ts'
 import get from './get.ts'
 import list from './list.ts'
 import set from './set.ts'
@@ -11,6 +12,7 @@ export default defineCommand({
   meta: { description: 'Variable operations' },
   subCommands: {
     list,
+    find,
     get,
     create,
     set,